requirements.txt
selenium
webdriver-manager
undetected-chromedriver
chromedriver-autoinstaller
PyMuPDF==1.23.8
beautifulsoup4
nltk
scikit-learn
spacy
numpy
pandas
Flask
google-cloud-storage
pyarrow
fastparquet
nl_core_news_sm @ https://github.com/explosion/spacy-models/releases/download/nl_core_news_sm-3.8.0/nl_core_news_sm-3.8.0-py3-none-any.whl
